# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Disbursement API",
        "contact": {
            "name": "Disbursement Team",
            "email": "support@email.com"
        },
        "version": "1.0"
    },
    "paths": {
        "/deposit/balance/{seller_id}": {
            "get": {
                "tags": [
                    "deposit"
                ],
                "operationId": "/deposit/balance/seller_id",
                "parameters": [
                    {
                        "name": "seller_id",
                        "in": "path",
                        "description": "Seller id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns seller deposit balance",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "status_code": {
                                            "type": "integer"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/balanceinfo"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Error: Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "status_code": {
                                            "type": "integer"
                                        },
                                        "seller_id": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/emptyarray"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/deposit/log": {
            "get": {
                "tags": [
                    "deposit"
                ],
                "operationId": "/deposit/log",
                "parameters": [
                    {
                        "name": "seller_id",
                        "in": "query",
                        "description": "Seller id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "$ref": "#/components/parameters/get_start_request_parameter"
                    },
                    {
                        "$ref": "#/components/parameters/get_limit_request_parameter"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns seller deposit log",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "status_code": {
                                            "type": "integer"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/balancelog"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Error: Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "status_code": {
                                            "type": "integer"
                                        },
                                        "seller_id": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/emptyarray"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/deposit/register": {
            "post": {
                "tags": [
                    "deposit"
                ],
                "operationId": "/deposit/register",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "seller_id": {
                                        "type": "string"
                                    },
                                    "amount": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Returns succes register seller deposit balance",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "status_code": {
                                            "type": "integer"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/balanceinforeg"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "status_code": {
                                            "type": "integer"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/emptyarray"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Error: Conflict",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "status_code": {
                                            "type": "integer"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/balanceinforeg"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "balanceinfo": {
                "properties": {
                    "balance": {
                        "type": "integer"
                    },
                    "seller_id": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "emptyarray": {},
            "balancelog": {
                "properties": {
                    "log_description": {
                        "type": "string"
                    },
                    "deposit_before": {
                        "type": "integer"
                    },
                    "amount": {
                        "type": "integer"
                    },
                    "updated_date": {
                        "type": "string"
                    },
                    "updated_by": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "balanceinforeg": {
                "properties": {
                    "amount": {
                        "type": "integer"
                    },
                    "seller_id": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "parameters": {
            "get_limit_request_parameter": {
                "name": "limit",
                "in": "query",
                "description": "Limit the number of results",
                "schema": {
                    "type": "number",
                    "default": 10
                }
            },
            "get_start_request_parameter": {
                "name": "start",
                "in": "query",
                "description": "Start from",
                "schema": {
                    "type": "number",
                    "default": 0
                }
            }
        }
    }
}